Parida mostrar apenas a lista para todos utilizadores

+++ gestor/listas_menu/index.php
@@ -1,3 +1,5 @@
+<?php $acesso = isset($_SESSION["tipo_acesso_logado"]) ? $_SESSION["tipo_acesso_logado"] : ""; ?>
+
 <ul class="navbar-nav me-auto mb-2 mb-lg-0">
 
     <li class="nav-item">
@@ -23,10 +25,12 @@
             Equipa
         </a>
         <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
+        <?php if( $acesso == "gestor" || $acesso == "administrador" ){ ?>
             <li><a class="dropdown-item" href="../equipa/cadastrar.php">Cadastrar</a></li>
             <li><a class="dropdown-item" href="../equipa/actualizar.php">Actualizar</a></li>
-            <li><a class="dropdown-item" href="../equipa/listar.php">Listar</a></li>
             <li><a class="dropdown-item" href="../equipa/eliminar.php">Eliminar</a></li>
+        <?php } ?>
+            <li><a class="dropdown-item" href="../equipa/listar.php">Listar</a></li>
         </ul>
     </li>
 
@@ -36,10 +40,12 @@
             Partida
         </a>
         <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
+        <?php if( $acesso == "gestor" || $acesso == "administrador" ){ ?>
             <li><a class="dropdown-item" href="../partida/cadastrar.php">Criar</a></li>
             <li><a class="dropdown-item" href="../partida/actualizar.php">Actualizar</a></li>
-            <li><a class="dropdown-item" href="../partida/listar.php">Listar</a></li>
             <li><a class="dropdown-item" href="../partida/eliminar.php">Eliminar</a></li>
+        <?php } ?>
+            <li><a class="dropdown-item" href="../partida/listar.php">Listar</a></li>
         </ul>
     </li>
